alarm-server/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
httpx==0.24.0
python-dotenv==1.0.0
langchain==0.0.174
langchain-google-genai==0.0.2
pytest==7.3.1
requests==2.31.0